+++ src/prefect/utilities/context.py
@@ -59,7 +59,7 @@ import threading
 from typing import Any, Iterator, MutableMapping
 
 from prefect.configuration import config, Config
-from prefect.utilities.collections import DotDict, as_nested_dict, merge_dicts
+from prefect.utilities.collections import DotDict, merge_dicts
 
 
 class Context(DotDict, threading.local):
@@ -78,11 +78,7 @@ class Context(DotDict, threading.local):
         super().__init__(*args, **kwargs)
         if "context" in config:
             self.update(config.context)
-        if "config" in self:
-            new_config = merge_dicts(config, self["config"])  # order matters
-            self["config"] = as_nested_dict(new_config, dct_class=Config)
-        else:
-            self["config"] = config
+        self["config"] = merge_dicts(config, self.get("config", {}))  # order matters
 
     def __repr__(self) -> str:
         return "<Context>"
@@ -101,8 +97,9 @@ class Context(DotDict, threading.local):
         try:
             new_context = dict(*args, **kwargs)
             if "config" in new_context:
-                new_config = merge_dicts(self.get("config", {}), new_context["config"])
-                new_context["config"] = as_nested_dict(new_config, dct_class=Config)
+                new_context["config"] = merge_dicts(
+                    self.get("config", {}), new_context["config"]
+                )
             self.update(new_context)  # type: ignore
             yield self
         finally: